mcp-servers/google_bot/requirements.txt
mcp>=1.12.0
google-api-python-client>=2.169.0
google-auth>=2.40.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
structlog>=23.2.0
python-dotenv>=1.0.0
